pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tally-db-pipeline"
version = "0.1.0"
description = "Standalone local pipeline for extracting Tally data into a queryable database."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "sqlalchemy>=2.0.36",
  "typer>=0.12.5",
]

[project.scripts]
tally-db-pipeline = "tally_db_pipeline.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
